+++ llarp/service/hidden_service_address_lookup.hpp
@@ -13,7 +13,7 @@ namespace llarp
     struct HiddenServiceAddressLookup : public IServiceLookup
     {
       const PubKey rootkey;
-      uint32_t relayOrder;
+      uint64_t relayOrder;
       const dht::Key_t location;
       using HandlerFunc = std::function< bool(
           const Address&, absl::optional< const IntroSet >, const RouterID&) >;
@@ -21,7 +21,7 @@ namespace llarp
 
       HiddenServiceAddressLookup(Endpoint* p, HandlerFunc h,
                                  const dht::Key_t& location,
-                                 const PubKey& rootkey, uint32_t relayOrder,
+                                 const PubKey& rootkey, uint64_t relayOrder,
                                  uint64_t tx);
 
       ~HiddenServiceAddressLookup() override = default;
